requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
ecdsa==0.18.0
idna==3.3
pycryptodome==3.15.0
requests==2.28.1
six==1.16.0
urllib3==1.26.10
